+++ dashboard/src/components/views/EnvironmentView/EntitiesContent.js
@@ -48,11 +48,13 @@ class EntitiesContent extends React.PureComponent {
     const { filter, limit, offset, order } = queryParams;
     const {
       data: { namespace } = {},
-      loading,
+      networkStatus,
       aborted,
       refetch,
-      poller,
     } = renderProps;
+
+    // see: https://github.com/apollographql/apollo-client/blob/master/packages/apollo-client/src/core/networkStatus.ts
+    const loading = networkStatus < 6;
 
     if (!namespace && !loading && !aborted) {
       return <NotFound />;
@@ -75,9 +77,7 @@ class EntitiesContent extends React.PureComponent {
                 editable={width !== "xs"}
                 limit={limit}
                 offset={offset}
-                loading={
-                  (loading && (!namespace || !poller.isRunning())) || aborted
-                }
+                loading={(loading && !namespace) || aborted}
                 onChangeQuery={setQueryParams}
                 namespace={namespace}
                 refetch={refetch}